requirements.txt
Flask==3.1.0
Werkzeug==3.1.3
pytest==8.3.4
requests==2.31.0
mysql-connector-python==8.0.32
api
mock
python-dotenv
mysqlclient
